requirements.txt
spotipy>=2.25.1
streamlit>=1.25.0
python-dotenv>=1.0.0
pandas>=2.0
streamlit-aggrid>=0.4.0
plotly>=5.0
